+++ fw_sim7600/sim7600/_dbus_descs.py
@@ -1,0 +1,141 @@
+#!/usr/bin/python3
+
+# List of supported DBus objects' definitions
+# Strings used as default value to populate the PID dict
+
+DEV_DBUS_DESC_SIM7600 = '''<node>
+  <interface name='{dbus_iface}'>
+    <property name="manufacturer" type="s" access="read">
+      <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal"
+       value="true"/>
+    </property>
+    <property name="model" type="s" access="read">
+      <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal"
+       value="true"/>
+    </property>
+    <property name="serial_number" type="s" access="read">
+      <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal"
+       value="true"/>
+    </property>
+    <property name="version_module" type="s" access="read">
+      <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal"
+       value="true"/>
+    </property>
+    <property name="version_chip" type="s" access="read">
+      <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal"
+       value="true"/>
+    </property>
+    <property name="version_firmware" type="s" access="read">
+      <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal"
+       value="true"/>
+    </property>
+    <property name="network_signal_quality_rssi" type="d" access="read">
+      <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal"
+       value="true"/>
+    </property>
+    <property name="network_signal_quality_ber" type="d" access="read">
+      <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal"
+       value="true"/>
+    </property>
+    <property name="network_sim_status" type="s" access="read">
+      <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal"
+       value="true"/>
+    </property>
+    
+    <property name="pos_gps_lat_dir" type="s" access="read">
+      <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal"
+       value="true"/>
+    </property>
+    <property name="pos_gps_lat_degrees" type="d" access="read">
+      <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal"
+       value="true"/>
+    </property>
+    <property name="pos_gps_log_dir" type="s" access="read">
+      <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal"
+       value="true"/>
+    </property>
+    <property name="pos_gps_log_degrees" type="d" access="read">
+      <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal"
+       value="true"/>
+    </property>
+    <property name="pos_gps_alt" type="d" access="read">
+      <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal"
+       value="true"/>
+    </property>
+    <property name="pos_gps_speed" type="d" access="read">
+      <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal"
+       value="true"/>
+    </property>
+    <property name="pos_gps_course" type="d" access="read">
+      <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal"
+       value="true"/>
+    </property>
+    
+    <property name="pos_gnss_lat_dir" type="s" access="read">
+      <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal"
+       value="true"/>
+    </property>
+    <property name="pos_gnss_lat_degrees" type="d" access="read">
+      <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal"
+       value="true"/>
+    </property>
+    <property name="pos_gnss_log_dir" type="s" access="read">
+      <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal"
+       value="true"/>
+    </property>
+    <property name="pos_gnss_log_degrees" type="d" access="read">
+      <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal"
+       value="true"/>
+    </property>
+    <property name="pos_gnss_alt" type="d" access="read">
+      <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal"
+       value="true"/>
+    </property>
+    <property name="pos_gnss_speed" type="d" access="read">
+      <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal"
+       value="true"/>
+    </property>
+    <property name="pos_gnss_course" type="d" access="read">
+      <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal"
+       value="true"/>
+    </property>
+    <property name="pos_gnss_pdop" type="d" access="read">
+      <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal"
+       value="true"/>
+    </property>
+    <property name="pos_gnss_hdop" type="d" access="read">
+      <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal"
+       value="true"/>
+    </property>
+    <property name="pos_gnss_vdop" type="d" access="read">
+      <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal"
+       value="true"/>
+    </property>
+    <property name="pos_gnss_mode" type="d" access="read">
+      <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal"
+       value="true"/>
+    </property>
+    <property name="pos_gnss_sat_gps_count" type="d" access="read">
+      <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal"
+       value="true"/>
+    </property>
+    <property name="pos_gnss_sat_glonass_count" type="d" access="read">
+      <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal"
+       value="true"/>
+    </property>
+    <property name="pos_gnss_sat_beidou_count" type="d" access="read">
+      <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal"
+       value="true"/>
+    </property>
+    
+    <property name="power_module_state" type="b" access="read">
+      <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal"
+       value="true"/>
+    </property>
+    <method name="power_module">
+      <arg direction="in" name="value" type="b"/>
+    </method>
+    
+  </interface>
+</node>
+'''