+++ CHANGELOG.md
@@ -1,8 +1,9 @@
-# FW UPS Pack V3 - Changelog
+# FW SIM 7600 - Changelog
 
 [README](README.md) | [CHANGELOG](CHANGELOG.md) | [TODOs](TODOs.md) | [LICENCE](LICENCE.md)
 
 
 ## Version 1.0.0
 
-* Adapted and refactored [com.robypomper.smartvan.fw.victron](https://github.com/Smart-Van-2-0/com.robypomper.smartvan.fw.victron) repo
+* Copied files from the [com.robypomper.smartvan.fw.upspack_v3](https://github.com/Smart-Van-2-0/com.robypomper.smartvan.fw.upspack_v3) repo
+* Implemented Device class for SIM 7600 over serial (UART)